+++ docs/changelog.rst
@@ -15,6 +15,8 @@ in specific versions.
 
 - Added :attr:`Player.is_playing` to check whether the player is currently
   playing a track (a track is loaded and the player is not paused).
+- Added :attr:`Player.remaining` to check the estimated playback time
+  remaining for the current track, calculated as ``track.length - player.position``.
 
 **Fixed**
 ~~~~~~~~~